// src/Api/national/dto/national.dto.ts
export class NationalApiDto {
  id: string;
  name: string;
  fullname: string;
  father: string;
  sex: string;
  marriage: string;
  birthdate: number;
  place: string;
  visatype: string;
  passport: string;
  amayesh: string;
  residencetype: string;
  otherresidencetype: string;
  visaexpire: number;
  economiccode: string;
  job: string;
  education: string;
  familymembersno: string;
  tel: string;
  mobile: string;
  sheba: string;
  familycode: string;
  address: string;
  postalcode: string;
  ref: string;
}